+++ ehpcinstant-20230701/ChangeLog.md
@@ -1,3 +1,14 @@
+2024-05-15 Version: 1.0.2
+- Update API CreateJob: update param Tasks.
+- Update API CreateJob: update response param.
+- Update API GetImage: add param ImageCategory.
+- Update API GetImage: add param ImageType.
+- Update API GetImage: update response param.
+- Update API ListImages: add param ImageCategory.
+- Update API ListImages: add param ImageType.
+- Update API ListImages: update response param.
+
+
 2024-05-11 Version: 1.0.1
 - Update API CreateJob: update response param.